(part 2 WIP) completes 2018/06 part 1
adds README to puzzle 2018/06
adds input.txt to puzzle 2018/06

Stuck on part 2. The example input on the site works fine but the actual answer of the real input is incorrect. I'm now stuck with 32268 as my answer for now which is too low. Debugging my sloppy code seems like hell so I might just rewrite it (for the 3rd time btw).

If i'm going to rewrite this I want to start out with getting the types out of the away before i write anything in the main. Furthermore I should not rush simple solutions, but instead try to go for more elegant and readable ones.